Added exceptions for courses + minor fixes

+++ Backend/src/main/java/com/example/springboot/services/CourseServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.springboot.services;
 
+import com.example.springboot.exceptions.NoSuchCourseExistsException;
+import com.example.springboot.exceptions.NoSuchUserExistsException;
 import com.example.springboot.models.Course;
+import com.example.springboot.models.User;
 import com.example.springboot.repositories.CourseRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,13 +18,13 @@ public class CourseServiceImpl implements CourseService {
     @Autowired
     private CourseRepository courseRepository;
 
-    public Course findById(@PathVariable Long courseId) throws Exception {
+    public Course findById(@PathVariable Long courseId) {
         return courseRepository.findById(courseId)
                 .orElseThrow(()
-                        -> new Exception("No course found with id = " + courseId));
+                        -> new NoSuchCourseExistsException("No course found with id = " + courseId));
     }
 
-    public List<Course> filterByName(@PathVariable String courseName) throws Exception {
+    public List<Course> filterByName(@PathVariable String courseName) {
         return courseRepository.findAll().stream().
                 filter(x -> x.getName().contains(courseName)).
                 collect(Collectors.toList());
@@ -43,11 +46,23 @@ public class CourseServiceImpl implements CourseService {
     }
 
     public String update(Course course) {
+        Optional<Course> courseToUpdate = courseRepository.findById(course.getId());
+
+        if (courseToUpdate.isEmpty()) {
+            throw new NoSuchCourseExistsException("No course found with id = " + course.getId());
+        }
+
         courseRepository.save(course);
         return "Course updated successfully";
     }
 
     public String delete(Long courseId) {
+        Optional<Course> courseToDelete = courseRepository.findById(courseId);
+
+        if (courseToDelete.isEmpty()) {
+            throw new NoSuchCourseExistsException("No course found with id = " + courseId);
+        }
+
         courseRepository.deleteById(courseId);
         return "Course deleted successfully";
     }
